Elevator changes robotcontainer changes other changes

+++ src/main/java/frc/robot/constants/Constants.java
@@ -26,36 +26,36 @@ public final class Constants {
 
     // public static final int DRIVETRAIN_PIGEON_ID = 0; // FIXME Set Pigeon ID
     // needs fixed
-    public static final double FRONT_LEFT_MODULE_DRIVE_MOTOR =   (6);
-    public static final double FRONT_RIGHT_MODULE_DRIVE_MOTOR =   (7);
-    public static final double BACK_LEFT_MODULE_DRIVE_MOTOR =   (9);
-    public static final double BACK_RIGHT_MODULE_DRIVE_MOTOR =   (8);
+    public static final int FRONT_LEFT_MODULE_DRIVE_MOTOR =   (6);
+    public static final int FRONT_RIGHT_MODULE_DRIVE_MOTOR =   (7);
+    public static final int BACK_LEFT_MODULE_DRIVE_MOTOR =   (9);
+    public static final int BACK_RIGHT_MODULE_DRIVE_MOTOR =   (8);
 
-    public static final double FRONT_LEFT_MODULE_STEER_MOTOR =   (10);
-    public static final double FRONT_RIGHT_MODULE_STEER_MOTOR =   (11);
-    public static final double BACK_LEFT_MODULE_STEER_MOTOR =   (13);
-    public static final double BACK_RIGHT_MODULE_STEER_MOTOR =   (12);
+    public static final int FRONT_LEFT_MODULE_STEER_MOTOR =   (10);
+    public static final int FRONT_RIGHT_MODULE_STEER_MOTOR =   (11);
+    public static final int BACK_LEFT_MODULE_STEER_MOTOR =   (13);
+    public static final int BACK_RIGHT_MODULE_STEER_MOTOR =   (12);
 
-    public static final double FRONT_LEFT_MODULE_STEER_ENCODER =   (15);
-    public static final double FRONT_RIGHT_MODULE_STEER_ENCODER =   (16);
-    public static final double BACK_LEFT_MODULE_STEER_ENCODER =   (18);
-    public static final double BACK_RIGHT_MODULE_STEER_ENCODER =   (17);
+    public static final int FRONT_LEFT_MODULE_STEER_ENCODER =   (15);
+    public static final int FRONT_RIGHT_MODULE_STEER_ENCODER =   (16);
+    public static final int BACK_LEFT_MODULE_STEER_ENCODER =   (18);
+    public static final int BACK_RIGHT_MODULE_STEER_ENCODER =   (17);
 
-    public static final double BACK_INTAKE_MOTOR =  (21);
-    public static final double SIDE_INTAKE_MOTOR =  (22);
-    public static final double INTAKE_OPENER_MOTOR =  (23);
+    public static final int BACK_INTAKE_MOTOR =  (21);
+    public static final int SIDE_INTAKE_MOTOR =  (22);
+    public static final int INTAKE_OPENER_MOTOR =  (23);
 
     // public static final doublePIVOTY_THRUBORE_ENCODER =  (31);
-    public static final double PIVOTY_MOTOR =  (32);
+    public static final int PIVOTY_MOTOR =  (32);
 
     // public static final doubleELEVEATOR_ENCODER =  (41);
-    public static final double ELEVATOR_MOTOR =  (42);
+    public static final int ELEVATOR_MOTOR =  (42);
 
     public static final int MIN_EXTEND_BREAK_BEAM = 0;
     public static final int MAX_EXTEND_BREAK_BEAM = 2;
     public static final int MID_EXTEND_BREAK_BEAM = 4;
     public static final int PIVOTY_BREAK_BEAM = 6;
-    public static final double COLOR_SENSOR = 8;
+    public static final int COLOR_SENSOR = 8;
     // In degrees
     // increasing turns clockwise
     public static final double FRONT_LEFT_MODULE_STEER_OFFSET = Math.toRadians(-11.46);
